Fix bug in moves.

+++ scala/model/src/main/scala/fr/iscpif/spacematters/model/initial/package.scala
@@ -30,7 +30,7 @@ package object initial {
 
     for {
       Vector(i, j, c) ← capacities
-    } matrix(i.toInt)(j.toInt) = c.toInt
+    } matrix(i.toInt)(j.toInt) = c.toDouble.toInt
 
     matrix.map(_.toSeq).toSeq
   }
